Updated products stock functionality

+++ src/actions/product.js
@@ -1,3 +1,4 @@
+import Swal from "sweetalert2";
 import { fetchSinToken } from "../helpers/fetch";
 import { types } from "../types/types";
 
@@ -6,27 +7,10 @@ export const productStartLoadingData = () => {
 
 		try {
 
-			const resp = await fetchSinToken('product');
+			const resp = await fetchSinToken('product/');
 			const body = await resp.json();
 
 			dispatch(productSetData(body))
-
-		} catch (error) {
-			console.log(error);
-		}
-
-	}
-}
-
-export const productStartLoad = (id) => {
-	return async (dispatch) => {
-
-		try {
-
-			const resp = await fetchSinToken(`product/${id}/`);
-			const body = await resp.json();
-
-			dispatch(productSetActive(body))
 
 		} catch (error) {
 			console.log(error);
@@ -40,9 +24,18 @@ export const productStartCreate = (product, router) => {
 		try {
 
 			const resp = await fetchSinToken('product/', product, 'POST');
-			const body = await resp.json();
-			dispatch(productAddNew(body))
-			router.push('/products')
+			if (resp.ok) {
+				const body = await resp.json();
+				dispatch(productAddNew(body))
+				Swal.fire(
+					'Creado!',
+					'Se creo el producto con exito.',
+					'success'
+				)
+				router.push('/products')
+			} else {
+				Swal.fire('Error', 'Algo salio mal, vuelva a intentar.', 'error')
+			}
 
 		} catch (error) {
 			console.log(error);
@@ -55,36 +48,62 @@ export const productStartUpdate = (product, router) => {
 
 		try {
 			const resp = await fetchSinToken(`product/${product.id}/`, product, 'PUT');
-			const body = await resp.json();
-
-			dispatch(productAddNew(body))
-			router.push('/products')
+			if (resp.ok) {
+				const body = await resp.json();
+				dispatch(productUpdated(body))
+				Swal.fire(
+					'Actualizado!',
+					'Se actualizado el producto con exito.',
+					'success'
+				)
+				router.push('/products')
+			} else {
+				Swal.fire('Error', 'Algo salio mal, vuelva a intentar.', 'error')
+			}
 
 		} catch (error) {
 			console.log(error)
 		}
-
 	}
 }
 
 export const productStartDelete = (id, router) => {
-	return async (dispatch, getState) => {
-		// const { product } = getState().productActive
+	return async (dispatch) => {
 		try {
 
-			await fetchSinToken(`product/${id}/`, '', 'DELETE');
-			dispatch(productDeleted(id))
+			Swal.fire({
+				title: '¿Estas seguro?',
+				text: 'Al eliminar este Producto se eliminaran todas las referencias que tiene en el sistema. Se recomienda archivar el producto.',
+				icon: 'warning',
+				confirmButtonColor: '#d33',
+				confirmButtonText: 'Entendido',
+				focusConfirm: false,
+				focusCancel: true,
+				showCancelButton: true,
+			}).then(async (result) => {
 
+				if (result.isConfirmed) {
+					const resp = await fetchSinToken(`product/${id}/`, '', 'DELETE');
+					if (resp.ok) {
+						dispatch(productDeleted(id))
+						Swal.fire(
+							'Eliminado!',
+							'Se ha eliminado con exito.',
+							'success'
+						)
+					} else {
+						Swal.fire('Error', 'Algo salio mal, vuelva a intentar.', 'error')
+					}
+				}
+
+			})
 		} catch (error) {
 			console.log(error);
 		}
 	}
 }
 
-export const productAddNew = (product) => ({
-	type: types.productAddNew,
-	payload: product
-})
+
 
 export const productSetData = (products) => ({
 	type: types.productSetData,
@@ -93,6 +112,16 @@ export const productSetData = (products) => ({
 
 export const productClearData = () => ({
 	type: types.productClearData,
+})
+
+export const productAddNew = (product) => ({
+	type: types.productAddNew,
+	payload: product
+})
+
+export const productUpdated = (product) => ({
+	type: types.productUpdated,
+	payload: product
 })
 
 export const productDeleted = (id) => ({
